+++ Use-Git.ps1
@@ -6,18 +6,18 @@
     .DESCRIPTION
         Calls the git application, with whatever arguments are provided.
 
-        Arguments can be provided with -GitArgument, which will automatically be bound to all parameters provided without a name.        
+        Arguments can be provided with -GitArgument, which will automatically be bound to all parameters provided without a name.
 
         Input can also be piped in.
-        
+
         If the input is a directory, Use-Git will Push-Location that directory.
         Otherwise, it will be passed as a positional argument (after any other arguments).
 
         Use-Git will combine errors and output, so that git output to standard error is handled without difficulty.
-    .NOTES        
+    .NOTES
         For almost everything git does, calling Use-Git is the same as calling git directly.
 
-        If you have any difficulties passing parameters to git, try enclosing them in quotes. 
+        If you have any difficulties passing parameters to git, try enclosing them in quotes.
     .LINK
         Out-Git
     .Example
@@ -69,17 +69,17 @@
         $callingContext =
             if ($callstackPeek.InvocationInfo.MyCommand.ScriptBlock) {
                 @($callstackPeek.InvocationInfo.MyCommand.ScriptBlock.Ast.FindAll({
-                    param($ast) 
+                    param($ast)
                         $ast.Extent.StartLineNumber -eq $myInv.ScriptLineNumber -and
-                        $ast.Extent.StartColumnNumber -eq $myInv.OffsetInLine -and 
+                        $ast.Extent.StartColumnNumber -eq $myInv.OffsetInLine -and
                         $ast -is [Management.Automation.Language.CommandAst]
                 },$true))[0]
             }
-        
+
         $argumentNumber = 0
 
         $gitArgsArray = [Collections.ArrayList]::new($GitArgument)
-        
+
         foreach ($commandElement in $callingContext.CommandElements) {
             if ($commandElement.parameterName -in 'd', 'v', 'c') {
                 # Insert the argument into the list
@@ -87,9 +87,11 @@
                     $argumentNumber - 1, # ( don't forget to subtract one, because the command is an element)
                     $commandElement.Extent.ToString()
                 )
-                if ($commandElement.parameterName -in 'd', 'c') { 
-                    $ConfirmPreference ='none' # so set confirm impact to none
-                }                
+                if ($commandElement.parameterName -in 'd', 'c', 'v') {
+                    $ConfirmPreference = 'none' # so set confirm impact to none
+                    $VerbosePreference = 'silentlyContinue'
+                    $DebugPreference   = 'silentlyContinue'
+                }
             }
             $argumentNumber++
         }
@@ -97,7 +99,7 @@
         $GitArgument = $gitArgsArray.ToArray()
 
         $progId = Get-Random
-        
+
         # A small number of git operations do not require a repo.  List them here.
         $RepoNotRequired = 'clone','init','version','help','-C'
 
@@ -116,8 +118,8 @@
         $directories = @()
         # Next, we need to create a collection of input object from each directory.
         $InputDirectories = [Ordered]@{}
-         
-         
+
+
         if (
             $AllInputObjects.Length -eq 0 -and # If we had no input objects and
             $myInv.PipelinePosition -gt 1 # are not the first step in the pipeline,
@@ -137,16 +139,16 @@
 
                     if ($directories) {      # If we have directories
                         # Store this file in the input object by each directory.
-                        $InputDirectories[$directories[-1]] = 
+                        $InputDirectories[$directories[-1]] =
                             # by forcing an existing entry into a list
-                            @($InputDirectories[$directories[-1]]) + 
+                            @($InputDirectories[$directories[-1]]) +
                             $in.Fullname # and adding this file name.
                     }
                 } elseif ($in -is [IO.DirectoryInfo]) {
                     $directories += Get-Item -LiteralPath $in.Fullname # If the input was a directory, keep track of it.
                 } else {
                     # Otherwise, see if it was a path and it was a directory
-                    if ((Test-Path $in -ErrorAction SilentlyContinue) -and 
+                    if ((Test-Path $in -ErrorAction SilentlyContinue) -and
                         (Get-Item $in -ErrorAction SilentlyContinue) -is [IO.DirectoryInfo]) {
                         $directories += Get-Item $in
                     } else {
@@ -158,11 +160,11 @@
                         }
                         if ($directories) {      # If we have directories
                             # Store this file in the input object by each directory.
-                            $InputDirectories[$directories[-1]] = 
+                            $InputDirectories[$directories[-1]] =
                                 # by forcing an existing entry into a list
-                                @($InputDirectories[$directories[-1]]) + 
+                                @($InputDirectories[$directories[-1]]) +
                                 $in # and adding this item.
-                        }                        
+                        }
                     }
                 }
             })
@@ -176,10 +178,10 @@
         if ($GitArgument -contains '-c' -or $GitArgument -contains '-C') {
             $paramCopy.Remove('Confirm')
         }
-        
+
         # Now, we will force there to be at least one directory (the current path).
         # This makes the code simpler, because we are always going thru a loop.
-        if (-not $directories) {            
+        if (-not $directories) {
             if ($GitArgument -ccontains '-C') {
                 $directories = $gitArgument[$GitArgument.IndexOf('-C') + 1]
             } else {
@@ -195,10 +197,10 @@
                 } else {
                     $dir
                 }
-            }) 
+            })
         }
 
-        
+
         $dirCount, $dirTotal = 0, $AllInputObjects.Length
 
         # For each directory we know of, we
@@ -216,30 +218,30 @@
                 $AllGitArgs = @($AllGitArgs -ne '')             # (skipping any empty arguments)
                 $OutGitParams = @{GitArgument=$AllGitArgs}      # and prepare a splat (to save precious space when reporting errors).
                 $dirCount++
-                
+
                 if ($WhatIfPreference) {
                     [ScriptBlock]::Create("git $($allGitArgs -join ' ')") |
                         Add-Member NoteProperty GitRoot $dir -Force -PassThru
                     Pop-Location
                     continue
                 }
-    
-    
+
+
                 if (-not $script:RepoRoots[$dir]) {             # and see if we have a repo root
-                    $script:RepoRoots[$dir] = 
+                    $script:RepoRoots[$dir] =
                         @("$(& $script:CachedGitCmd rev-parse --show-toplevel *>&1)") -like "*/*" -replace '/', [io.path]::DirectorySeparatorChar
                     if (-not $script:RepoRoots[$dir] -and      # If we did not have a repo root
-                        -not ($gitArgument -match "(?>$($RepoNotRequired -join '|'))") # and we are not doing an operation that does not require one 
+                        -not ($gitArgument -match "(?>$($RepoNotRequired -join '|'))") # and we are not doing an operation that does not require one
                     ) {
                         Write-Warning "'$($dir)' is not a git repository" # warn that there is no repo (#21)
-                        Pop-Location # pop back out of the directory                   
+                        Pop-Location # pop back out of the directory
                         continue nextDirectory # and continue to the next directory.
                     }
                 }
-                            
+
                 $OutGitParams.GitRoot = "$($script:RepoRoots[$dir])"
                 Write-Verbose "Calling git with $AllGitArgs"
-    
+
                 if ($dirCount -gt 1) {
                     # Clamp percentage complete within 0-100
                     $percentageComplete = [Math]::Max(
@@ -247,11 +249,11 @@
                             [Math]::Round(
                                 ([double]$dirCount / $dirTotal) * 100
                             ), 100
-                        ), 
+                        ),
                     0)
                     Write-Progress -PercentComplete $percentageComplete -Status "git $allGitArgs " -Activity "$($dir) " -Id $progId
                 }
-            
+
                 # If we have indicated we do not care about -Confirmation, don't prompt
                 if (($ConfirmPreference -eq 'None' -and (-not $paramCopy.Confirm)) -or
                     $PSCmdlet.ShouldProcess("$pwd : git $allGitArgs") # otherwise, as for confirmation to run.
@@ -268,7 +270,7 @@
                     & $script:CachedGitCmd @AllGitArgs *>&1       | # Then we run git, combining all streams into output.
                                                                     # then pipe to Out-Git, which will
                         Out-Git @OutGitParams # output git as objects.
-                                            
+
                         # These objects are decorated for the PowerShell Extended Type System.
                         # This makes them easy to extend and customize their display.
                         # If Out-Git finds one or more extensions to run, these can parse the output.
@@ -276,7 +278,7 @@
 
             }
 
-            
+
             Pop-Location # After we have run, Pop back out of the location.
         }
         if ($dirCount -gt 1) {