+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.3.7:
+
+* git remote
+ * Now supporting git remote! (Fixes #129)
+
+~~~PowerShell
+git remote | git remote show
+~~~
+
+Also, some improvements to the GitHub Action:
+
+* Icon Update (Fixes #132)
+* No longer using set-output (Fixes #131)
+* Adding -InstallModule to Action (Fixes #132)
+
+---
+
 ## 0.3.6:
 
 * git log